Implement minimal LangGraph content creation agent using OpenAI's ChatGPT
